[chore] - Application validation
